requirements.txt
amqp==5.3.1
asgiref==3.11.0
attrs==25.4.0
billiard==4.2.4
celery==5.5.3
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.3
defusedxml==0.7.1
dj-database-url==3.0.1
Django==5.2.8
django-cors-headers==4.9.0
django-filter==24.3
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
drf-spectacular==0.29.0
idna==3.11
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
oauthlib==3.3.1
packaging==25.0
pillow==12.0.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python3-openid==3.2.0
PyYAML==6.0.3
redis==7.0.1
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.30.0
six==1.17.0
social-auth-app-django==5.6.0
social-auth-core==4.8.1
sqlparse==0.5.4
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.6.2
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.11.0
